// packages/automl/frontend/src/app/types/pipeline.ts
export enum RuntimeStateKF {
  RUNTIME_STATE_UNSPECIFIED = 'RUNTIME_STATE_UNSPECIFIED',
  PENDING = 'PENDING',
  RUNNING = 'RUNNING',
  SUCCEEDED = 'SUCCEEDED',
  SKIPPED = 'SKIPPED',
  FAILED = 'FAILED',
  CANCELING = 'CANCELING',
  CANCELED = 'CANCELED',
  CACHED = 'CACHED',
  PAUSED = 'PAUSED',
}
// ...
export type TaskDetailKF = {
  run_id: string;
  task_id: string;
  display_name?: string;
  create_time: string;
  start_time: string;
  end_time: string;
  execution_id?: string;
  state?: RuntimeStateKF;
  state_history?: { update_time: string; state: string }[];
  child_tasks?: { pod_name: string; task_id?: string }[];
  parent_task_id?: string;
};

export type RunDetailsKF = {
  pipeline_context_id?: string;
  pipeline_run_context_id?: string;
  task_details: TaskDetailKF[];
};